Postpone autowrapping until next addchar() call.
Some programs will fill a row and then send CR+LF to go to the row
immediately below it. If autowrap acts as soon as we reach the end
of the row, it will result in  doubled CR+LF sequences, so we need
to make  sure the program  is actually  trying to add a  character
after the last column before we wrap the line.

An example of a program that needs this fix is procps-ng's top.