Write partial byte iif offset % 8 != 0.
The function end_key() flushes the  last non-empty sub-block in a block.
The length of  this sub-block depends on whether there  are bits waiting
to fill up a temporary byte. If gif->offset is a multiple of 8, then all
bits  were  already  packed  into  bytes  and  written  to  gif->buffer.
Otherwise, there are 8 - gif->offset % 8 bits that need to be written in
the buffer, adding one byte to the sub-block.