MIN/MAX delay controls
Make sure delay doesn't exceed 65535, the limit of the format.

Don't extend the last frame to MIN_DELAY, but don't use zero as some
player assume that's an error and extend it to the previous value.

GIF cannot represent a time interval before the first frame. Instead of
moving the script delay after the first frame move it after the last frame
so that it's still between the first and last frames when we're looping.